some kidding with intl and local

+++ app/helpers/cats_helper.rb
@@ -1,5 +1,5 @@
 module CatsHelper
   def localize_date(date)
-    I18n.l date, format: :short if date
+    I18n.l date, format: :long if date
   end
 end
